split callback test utils for better readability
