+++ sklearn/callback/tests/_common/__init__.py
@@ -1,0 +1,2 @@
+# Authors: The scikit-learn developers
+# SPDX-License-Identifier: BSD-3-Clause
