+++ sklearn/callback/tests/test_callback_context.py
@@ -12,16 +12,18 @@ from sklearn.callback._callback_context import (
     _from_reconstruction_attributes,
     get_context_path,
 )
-from sklearn.callback.tests._utils import (
+from sklearn.callback.tests._common.callbacks import (
+    NotRequiredKwargsCallback,
+    RecordingAutoPropagatedCallback,
+    RecordingCallback,
+    StopFitCallback,
+)
+from sklearn.callback.tests._common.estimators import (
     MaxIterEstimator,
     MetaEstimator,
     NoCallbackEstimator,
     NoSubtaskEstimator,
-    NotRequiredKwargsCallback,
     ParentFitEstimator,
-    RecordingAutoPropagatedCallback,
-    RecordingCallback,
-    StopFitCallback,
     ThirdPartyEstimator,
 )
